add globalstyles
