+++ src/components/AuthCardInfo/AuthCardInfo.tsx
@@ -1,25 +1,20 @@
 import { Box, Typography, Icon } from "@mui/material";
-import LogoSVG from "../../assets/logos/authLogo";
-import styled from "styled-components";
+
+import MainLogo from "../../assets/logos/authLogo";
+
+import { StyledBox } from "./styles";
 
 interface IAuthCardInfo {
   mainText: string;
   additionalText: string;
 }
 
-const StyledBox = styled(Box)`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-`;
-
 const AuthCardInfo = ({ mainText, additionalText }: IAuthCardInfo) => {
   return (
     <>
       <StyledBox>
         <Icon fontSize="small">
-          <LogoSVG />
+          <MainLogo />
         </Icon>
         <Typography variant="authLogoText">Dashboard Kit</Typography>
       </StyledBox>
