+++ src/components/AuthCardInfo/styles.tsx
@@ -1,0 +1,9 @@
+import { Box, Typography, Icon } from "@mui/material";
+import styled from "styled-components";
+
+export const StyledBox = styled(Box)`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 12px;
+`;
