+++ src/index.tsx
@@ -2,9 +2,14 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import { ThemeProvider } from "@mui/material/styles";
-import { StyledEngineProvider } from "@mui/material";
+import { StyledEngineProvider, GlobalStyles } from "@mui/material";
+
 import App from "./App";
 import theme from "./theme";
+
+const globalStyles = {
+  "*": { margin: 0, boxSizing: "border-box" },
+};
 
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
@@ -12,6 +17,7 @@ const root = ReactDOM.createRoot(
 root.render(
   <StyledEngineProvider injectFirst>
     <ThemeProvider theme={theme}>
+      <GlobalStyles styles={globalStyles} />
       <App />
     </ThemeProvider>
   </StyledEngineProvider>
