--- src/screens/SignUp/Styles.tsx
@@ -1,7 +1,0 @@
-export const containerStyles = {
-  backgroundColor: "#363740",
-
-  padding: "5%",
-  width: "100%",
-  height: "100%",
-};
